Changed zoom selector to input range slider.
The right tool bar is now gone and the canvas container has expanded.
The zoom buttons have changed to a slider in the footer. The max zoom
factor is determined when an image is opened and thus varies with
different images to prevent OOMing.

R=calamity@chromium.org, jackhou@chromium.org

Review URL: https://codereview.appspot.com/57470043