Keep tool operations active when mouse leaves the canvas.
This lets you start drawing, mouse off the canvas and mouse back again,
and have the draw operation continue.

To make this change the window now listens for mouse events, instead
of the canvas wrapper.

Another functional change is that the mouse coordinates are hidden when
the mouse is not inside the canvas.

The behavior when zoomed in could be refined but works pretty well as
is.

R=jackhou@chromium.org

Review URL: https://codereview.appspot.com/71020043