Fix bitmapper Makefile dependencies
This adds proper dependency tracking so that `make` doesn't do
unnecessary work when files are not changed.

It ensures make re-vulcanizes when any .html file changes under the app
dir.

It moves generic dependencies to a `requirements.mk` file to isolate
make rules that are not specific to bitmapper.

R=hbeder@google.com

Review URL: https://codereview.appspot.com/178610043