Change bitmapper icon color to red for debug version.
The background color of the bitmapper icon is changed to red, as described in
the bug link below. A folder icon/badge has been added which contains these
"badged" icons, and these icons will be used in the debug builds as reflected
in the Makefile.

BUG=https://github.com/jackhou-chromium/bitmapper/issues/29
R=tapted@chromium.org

Review URL: https://codereview.appspot.com/194780043