Allow canvas to be resized by dragging edges
Done by detecting when the mouse is on the edge corners
(width and height), and when it is at such corners,
the canvas will be able to be dragged to resize.
When doing so, the size of new canvas will be made visible
by adjusting the canvasPlaceholder to reflect the current
mouse co-ordinates.

BUG=https://github.com/jackhou-chromium/bitmapper/issues/34
R=tapted@chromium.org

Review URL: https://codereview.appspot.com/190440043