Fixed several issues relating to zoomFactor / zoomPercent.
Change formula for percentToFactor and factorToPercent, so that there
are no 'illegal' values as there was previously.

The definition of zoomFactor has been changed, such that values > 0,
represent 'zooming in' and values < 0, represent 'zooming out'. The
default zoom is 100% and the value will be 0 at this point.

The zoom percentage is no longer a stored variable, and is calculated
when it is being rendered to the user from the current zoom level
(which is a number greater than 0).

BUG=https://github.com/jackhou-chromium/bitmapper/issues/49
R=mgiuca@chromium.org

Review URL: https://codereview.appspot.com/189440043