Show canvas size during resize operation
The canvas size (Width and Height) in pixels will be displayed in the
bottom left corner, while the canvas is being resized by dragging the
edges.

The size is displayed in a canvasDimensions div, which by default will
have a visibility hidden. The div will only be made visible (by setting
visibility to visible) while the width or height is being resized.

BUG=https://github.com/jackhou-chromium/bitmapper/issues/51
R=mgiuca@chromium.org

Review URL: https://codereview.appspot.com/195730044