Allow crop-to-selection with use of selection tool.
A button will be added, upon selection of the selection tool, which
allow the functionality of 'crop to selection'. The button will only
be clickable when something on the canvas has been selected.

A new polymer element, using paper-icon-button has been created to allow
for this functionality. When the selection tool is not selected, the
button will have a display of none, and visibility hidden to ensure
that it will not get rendered onto the toolPanel.

The styling of the crop-to-selection button has been adopted from the
other paper-icon-button, previously present on the application. It will contain the crop icon, from the
image editor icon set.

BUG=https://github.com/jackhou-chromium/bitmapper/issues/53
R=mgiuca@chromium.org

Review URL: https://codereview.appspot.com/194020043