First stroke of brushStroke to use drawBrush instead of drawLine.
When using the brush tool, if the current mouse coordinates and the
last mouse coordinates do not differ, a single point (being in the form
of a circle should be drawn onto the canvas.

The current implementation of drawBrush uses a moveTo operation which
is followed by a lineTo operation to achieve the circular motion of the
brush. However, in the case that the last and previous points do not
differ (for example, the initial brush stroke) nothing will get drawn
onto the canvas. This is because for the lineTo operation to draw a
'brushStroke' it requires 2 different points.

When the current and last points do not differ, the circular brush
pattern will be implmented by drawing a circle onto the canvas, at
the current mouse point. This will be done be the arc method on
the canvas, and using a radius which is half the current
sizeSelector value. The opacity of the brushTool will be maintained by
setting the fill property to the current color, with opacity taken into
account.

BUG=https://github.com/jackhou-chromium/bitmapper/issues/59
R=mgiuca@chromium.org

Review URL: https://codereview.appspot.com/189530043