Fix problem with resizing by drag when using a tool on canvas edge.
The canvas can be accidentally resized by using a tool near the edge of
the canvas, as described in the bug.
It can be easily seen when attempting to draw from the edges of the
canvas.

The problem is due to event bubbling of the mouseDown event from the
canvasWrapper to the canvasViewport. As usage of the currently selected
tool, should not result in a resize operation occuring (which gets set
on the canvasWrapper event). By preventing the propagation of the
mouseDown event, the resizeOperation will not take place unecessarily.

BUG=https://github.com/jackhou-chromium/bitmapper/issues/57
R=mgiuca@chromium.org

Review URL: https://codereview.appspot.com/193020043