Allow alpha blending when using brush tool.
When the brush tool brushes over itself, the overlapping components
should not appear darker.

A new brushCanvas has been created, using the brushTool will draw onto
this canvas, which will then get captured as an image. This image is
what is displayed while using the brush tool (on a single mouse down
operation). The image will have the opacity style set to the current
opacity of the brush tool. On a mouse, the image will get drawn to the
canvas, with the current opacity being presevered by setting the global
alpha value.

BUG=https://github.com/jackhou-chromium/bitmapper/issues/58
R=mgiuca@chromium.org

Review URL: https://codereview.appspot.com/195890043