Allow proper alpha blending for eraserTool.
Initial implementation for alpha blending can be found at cl/195890043,
and for pencil and eraser tool cl/195960043.
The eraserTool now takes opacity into account, and has the capability
of making the brushed over components on the canvas 'lighter'.

The eraser is initially drawn onto the brushCanvas, as a black stroke
with opacity value 1. When the brushCanvas is drawn onto the
displayCanvas, the globalAlpha value will get set (as the same as
alpha blending on pencil and brush tools), but will also get
globalCompositeOperation set to 'destination-out'.

R=mgiuca@chromium.org

Review URL: https://codereview.appspot.com/197870043