Only change displayCanvas height if canvasViewport is defined.
Change resulted due to running of tests.
clientWidth and clientHeight are read only properties, and so when running tests these values can not be set.
An example test which depends on this is the pixelTransparency test in penciltooltest.

The current implementation means that clientHeight/Width is not defined when running the test suite.

R=hbeder@google.com, mgiuca@chromium.org

Review URL: https://codereview.appspot.com/196150043