Fix eraserTool functionality in release.
Similar problem with closure as in
https://codereview.appspot.com/199350043

Erasertool was not functioning correctly in release due to closure,
and the problem of bitmapper.selected.constructor.name in the
eraserToolSelected functio, which is used to check whether the eraser
tool is selected.

It wasn't possible to use constructor.name in a similar 'closure'
method, as in the preivous cl. So instead, I've accessed the
selectedtool through its polymer attributes.
BUG=https://github.com/jackhou-chromium/bitmapper/issues/77
R=hbeder@google.com, mgiuca@chromium.org

Review URL: https://codereview.appspot.com/202790043