Use image-rendering: pixelated with displayCanvas.
This changes the canvas size of the displayCanvas to be the size of the
visible portion of the image, rather than the size of the viewport. I.e.
it does not include zoom. Zoom is achieved by setting the layout size
to the size of the viewport, and adding image-rendering: pixelated to
get nearest neighbor scaling.

We cannot yet show the sourceCanvas and brushCanvas directly because
there is no way to blend them with 'destination-out'.

R=mgiuca@chromium.org

Review URL: https://codereview.appspot.com/220240043